Catch 'invalid ordStatus' on an amend.
It's possible, given latency over the internet, that an order fills or cancels
while the bot is trying to amend it. If that happens, the bot halts the current
tick, waits 500ms, and starts another. This ensures the new order status
has been reflected in the data from the websocket.